Update to 1.0.0 release

+++ libosinfo.spec
@@ -1,18 +1,8 @@
 # -*- rpm-spec -*-
 
-# Plugin isn't ready for real world use yet - it needs
-# a security audit at very least
-%define with_plugin 0
-
-%define with_gir 0
-
-%if 0%{?fedora} >= 15 || 0%{?rhel} >= 7
-%define with_gir 1
-%endif
-
 Summary: A library for managing OS information for virtualization
 Name: libosinfo
-Version: 0.3.1
+Version: 1.0.0
 Release: 1%{?dist}%{?extra_release}
 License: LGPLv2+
 Group: Development/Libraries
@@ -29,10 +19,10 @@ BuildRequires: vala-tools
 BuildRequires: libsoup-devel
 BuildRequires: /usr/bin/pod2man
 BuildRequires: hwdata
-%if %{with_gir}
 BuildRequires: gobject-introspection-devel
-%endif
 Requires: hwdata
+Requires: osinfo-db
+Requires: osinfo-db-tools
 
 %description
 libosinfo is a library that allows virtualization provisioning tools to
@@ -69,13 +59,7 @@ This package provides the Vala bindings for libosinfo library.
 %setup -q
 
 %build
-%if %{with_gir}
-%define gir_arg --enable-introspection=yes
-%else
-%define gir_arg --enable-introspection=no
-%endif
-
-%configure %{gir_arg} --enable-vala=yes
+%configure --enable-introspection=yes --enable-vala=yes
 %__make %{?_smp_mflags} V=1
 
 chmod a-x examples/*.js examples/*.py
@@ -89,7 +73,11 @@ rm -f %{buildroot}%{_libdir}/*.la
 %find_lang %{name}
 
 %check
-make check
+if ! make check
+then
+  cat test/test-suite.log || true
+  exit 1
+fi
 
 %clean
 rm -fr %{buildroot}
@@ -102,26 +90,13 @@ rm -fr %{buildroot}
 %defattr(-, root, root)
 %doc AUTHORS ChangeLog COPYING.LIB NEWS README
 %{_bindir}/osinfo-detect
-%{_bindir}/osinfo-db-validate
 %{_bindir}/osinfo-query
 %{_bindir}/osinfo-install-script
-%dir %{_datadir}/libosinfo/
-%dir %{_datadir}/libosinfo/db/
-%dir %{_datadir}/libosinfo/schemas/
-%{_datadir}/libosinfo/db/datamap
-%{_datadir}/libosinfo/db/device
-%{_datadir}/libosinfo/db/os
-%{_datadir}/libosinfo/db/platform
-%{_datadir}/libosinfo/db/install-script
-%{_datadir}/libosinfo/schemas/libosinfo.rng
-%{_mandir}/man1/osinfo-db-validate.1*
 %{_mandir}/man1/osinfo-detect.1*
 %{_mandir}/man1/osinfo-query.1*
 %{_mandir}/man1/osinfo-install-script.1*
 %{_libdir}/%{name}-1.0.so.*
-%if %{with_gir}
 %{_libdir}/girepository-1.0/Libosinfo-1.0.typelib
-%endif
 
 %files devel
 %defattr(-, root, root)
@@ -132,9 +107,7 @@ rm -fr %{buildroot}
 %dir %{_includedir}/%{name}-1.0/osinfo/
 %{_includedir}/%{name}-1.0/osinfo/*.h
 %{_libdir}/pkgconfig/%{name}-1.0.pc
-%if %{with_gir}
 %{_datadir}/gir-1.0/Libosinfo-1.0.gir
-%endif
 %{_datadir}/gtk-doc/html/Libosinfo
 
 %files vala
@@ -142,6 +115,9 @@ rm -fr %{buildroot}
 %{_datadir}/vala/vapi/libosinfo-1.0.vapi
 
 %changelog
+* Fri Oct  7 2016 Daniel P. Berrange <berrange@redhat.com> 1.0.0-1
+- New upstream release 1.0.0
+
 * Fri Jul  1 2016 Daniel P. Berrange <berrange@redhat.com> 0.3.1-1
 - New upstream release 0.3.1
